requirements.txt
openai
streamlit
langchain
urllib3==1.26.6
python-dotenv==1.0.0
deeplake
tiktoken
google-api-python-client
huggingface_hub